Struct refuse::AnyRef

pub struct AnyRef { /* private fields */ }

A type-erased garbage collected reference.

Implementations

impl AnyRef

pub const fn downcast_ref<T>(&self) -> Ref<T>
where T: Collectable,

Returns a Ref<T>.

This function does not do any type checking. If T is not the correct type, attempting to load the underyling value will fail.

pub fn upgrade(&self, guard: &CollectionGuard<'_>) -> Option<AnyRoot>

Returns a root for this reference, if the value has not been collected.

pub fn downcast_checked<T>(&self) -> Option<Ref<T>>
where T: Collectable,

Returns a Ref<T>, if T matches the type of this reference.

pub fn downcast_root<T>(&self, guard: &CollectionGuard<'_>) -> Option<Root<T>>
where T: Collectable,

Returns a Root<T> if the underlying reference points to a T that has not been collected.

pub fn load<'guard, T>( &self, guard: &'guard CollectionGuard<'_>, ) -> Option<&'guard T>
where T: Collectable,

Loads a reference to the underlying data. Returns None if the data has been collected and is no longer available.

pub fn load_mapped<'guard, T>( &self, guard: &'guard CollectionGuard<'_>, ) -> Option<&'guard T>
where T: ?Sized + 'static,

Returns a reference to the result of MapAs::map_as(), if the value has not been collected and MapAs::Target is T.