Struct refuse::Root

pub struct Root<T>
where
    T: Collectable,
{ /* private fields */ }

A root reference to a T that has been allocated in the garbage collector.

This type behaves very similarly to Arc<T>. It implements Deref<Target = T>, and it is also cheap-to-clone, utilizing atomic reference counting to track the number of root references currently exist to the underlying value.

While any root references exist for a given allocation, the garbage collector will not collect the allocation.

Implementations

impl<T> Root<T>
where T: Collectable,

pub fn new<'a>(value: T, guard: impl AsRef<CollectionGuard<'a>>) -> Self

Stores value in the garbage collector, returning a root reference to the data.

pub fn root_count(&self) -> u64

Returns the current number of root references to this value, including self.

pub const fn downgrade(&self) -> Ref<T>

Returns a “weak” reference to this root.

pub fn downgrade_any(&self) -> AnyRef

Returns an untyped “weak” reference erased to this root.

pub fn as_any(&self) -> AnyRef

Returns this reference as an untyped reference.

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if these two references point to the same underlying allocation.