Trait refuse::Trace

pub trait Trace: Send + Sync + 'static {
    const MAY_CONTAIN_REFERENCES: bool;

    // Required method
    fn trace(&self, tracer: &mut Tracer<'_>);
}

A type that can find and mark any references it has.

Required Associated Constants

const MAY_CONTAIN_REFERENCES: bool

If true, this type may contain references and should have its trace() function invoked during the collector’s “mark” phase.

Required Methods

fn trace(&self, tracer: &mut Tracer<'_>)

Traces all refrences that this value references.

This function should invoke Tracer::mark() for each Ref<T> it contains. Failing to do so will allow the garbage collector to free the data, preventing the ability to load() the data in the future.

Object Safety

This trait is not object safe.